% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banner.R
\name{banner}
\alias{banner}
\title{Banner Function}
\usage{
banner(Text = "", size = 1, color = "grey", font = 2, ...)
}
\arguments{
\item{Text}{character string; text to display inside banner}

\item{size}{numeric (1 stays the same, 1.2 increases size by 20% etc.); size of banner}

\item{color}{banner background defaults to grey}

\item{font}{defasults to bold (2)}

\item{...}{other text parameters including cex, etc.}
}
\value{
Banner/text on top of plot
}
\description{
Create ggplot style banners on the fly - addition to existing plot
}

