% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basechange.r
\name{base.change}
\alias{base.change}
\title{Base Change}
\usage{
base.change(x, base.origin, base.out)
}
\arguments{
\item{x}{a number; must be alphanumeric for base greater than 10}

\item{base.origin}{integer or character; 1-26 indicating base of x. Must a character value for alphanumeic numbers}

\item{base.out}{a number; 1-26 indicating base of returned x}
}
\value{
number in base of choice
}
\description{
Takes a number of any base 1-26 and returns the same number with the base of your choice (1-26)
}
\examples{
base.change("39F2",16,2) # hexadecimal to binary
paste(base.change(989,10,16),base.change(2002633725,10,16),sep=".") # PIT code 989.2002633725 from decimal to hexadecimal
}

