% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notin.r
\name{\%nin\%}
\alias{\%nin\%}
\title{Not In}
\usage{
x \%nin\% y
}
\arguments{
\item{x}{a vector; values to be nonmatched}

\item{y}{a vector; values to be nonmatched against}
}
\value{
boolean vector of nonmatches
}
\description{
Inverse of %in%. Returns a logical vector indicating nonmatches for its left operand.
}

