% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headtails.r
\name{ht}
\alias{ht}
\title{Head/Tails}
\usage{
ht(x)
}
\arguments{
\item{x}{data frame or matrix}
}
\value{
first and last 10 lines of data frame or matrix
}
\description{
displays first and last 10 lines of data frame or matrix
}

