% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.r
\name{scrape}
\alias{scrape}
\title{Web Scrape}
\usage{
scrape(url)
}
\arguments{
\item{url}{character string; url to webpage}
}
\value{
data.frame of web content
}
\description{
Quick and dirty function used to scrape html content from websites. Function may need to be modified depending on situational needs.
}

